make_synthesis_engineを移動 (#173)
* キャンセル機能を追加

* キャンセル可能なことを明示

* 型ヒントを追加

* グローバル変数をクラス化

* multiprocessing.freeze_support()の行を変更

https://docs.python.org/ja/3/library/multiprocessing.html#multiprocessing.freeze_support

* 変数名をリネーム

* 重複した処理を関数化

* デフォルトで機能を無効化

* 引数を追加

* ProcessManager->CancellableEngine

* docstringを追加

* pysen通るように修正

* 一旦別ファイルに切り出し

* fix pysen

* cancellable engineを削除

* importを修正
